削除機能の再修正

+++ app/controllers/items_controller.rb
@@ -35,7 +35,7 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if current_user.id == @item.user_id
+    if user_signed_in? && current_user.id == @item.user_id
       redirect_to root_path if @item.destroy
     end
   end
